Add AsyncImage ViewBuilder
