+++ CommonUI/Sources/FeatureComponent/MovieCard.swift
@@ -3,17 +3,14 @@ import DataModels
 import SwiftUI
 
 public struct MovieCard: View {
-    let url: URL
-    public init?(movie: Movie) {
-        guard let movieURL = URL(string: movie.posterUrl) else {
-            return nil
-        }
-        self.url = movieURL
+    let movie: Movie
+    public init(movie: Movie) {
+        self.movie = movie
     }
     
     public var body: some View {
         VStack {
-            AsyncImage(url: url) { phase in
+            AsyncImageViewBuilder(urlString:movie.posterUrl) { phase in
                 if let image = phase.image {
                     image
                         .resizable()
@@ -30,16 +27,6 @@ public struct MovieCard: View {
                 }
             }
             .padding(10)
-        }
-        .onAppear {
-            // Load and cache the image using URLSession
-            let request = URLRequest(url: url, cachePolicy: .returnCacheDataElseLoad)
-            URLSession.shared.dataTask(with: request) { data, response, _ in
-                if let data = data, let response = response {
-                    let cachedResponse = CachedURLResponse(response: response, data: data)
-                    URLCache.shared.storeCachedResponse(cachedResponse, for: request)
-                }
-            }.resume()
         }
     }
 }
