+++ FavoritesMicro/Sources/FavoritesMicro/FavoriteNavigationView/FavoriteNavigationViewModel.swift
@@ -4,7 +4,7 @@ import FeatureView
 class FavoriteNavigationViewModel {
     // MARK: - View Builders
     
-    func createMovieDetailsView(movie: Movie) -> MovieDetailsView? {
+    func createMovieDetailsView(movie: Movie) -> MovieDetailsView {
         return MovieDetailsView(movie: movie)
     }
 }
