+++ SearchMicro/Sources/SearchMicro/SearchNavigation/SearchNavigationViewModel.swift
@@ -5,7 +5,7 @@ import SwiftUI
 class SearchNavigationViewModel {
     // MARK: - View Builders
     
-    func createMovieDetailsView(movie: Movie) -> MovieDetailsView? {
+    func createMovieDetailsView(movie: Movie) -> MovieDetailsView {
         return MovieDetailsView(movie: movie)
     }
 }
